[#100] Sample does no longer require lifetime parameters; remove traits Subscribe and Payload
