+++ iceoryx2/src/sample.rs
@@ -33,42 +33,63 @@
 //!
 //! See also [`crate::sample::Sample`].
 
+use std::rc::Rc;
 use std::{fmt::Debug, ops::Deref};
 
-use crate::port::subscribe::internal::SubscribeMgmt;
+use iceoryx2_bb_log::{fatal_panic, warn};
+use iceoryx2_cal::shared_memory::SharedMemory;
+use iceoryx2_cal::zero_copy_connection::{PointerOffset, ZeroCopyReceiver, ZeroCopyReleaseError};
+
+use crate::port::details::publisher_connections::PublisherConnections;
+use crate::raw_sample::RawSample;
 use crate::service::header::publish_subscribe::Header;
-use crate::{payload::Payload, raw_sample::RawSample};
 
 /// It stores the payload and is acquired by the [`crate::port::subscriber::Subscriber`] whenever
 /// it receives new data from a [`crate::port::publisher::Publisher`] via
 /// [`crate::port::subscribe::Subscribe::receive()`].
 #[derive(Debug)]
-pub struct Sample<'subscriber, MessageType: Debug> {
-    pub(crate) subscriber: &'subscriber dyn SubscribeMgmt,
+pub struct Sample<MessageType: Debug, Service: crate::service::Service> {
+    pub(crate) publisher_connections: Rc<PublisherConnections<Service>>,
     pub(crate) ptr: RawSample<Header, MessageType>,
     pub(crate) channel_id: usize,
 }
 
-impl<MessageType: Debug> Deref for Sample<'_, MessageType> {
+impl<MessageType: Debug, Service: crate::service::Service> Deref for Sample<MessageType, Service> {
     type Target = MessageType;
     fn deref(&self) -> &Self::Target {
         self.ptr.as_data_ref()
     }
 }
 
-impl<MessageType: Debug> Drop for Sample<'_, MessageType> {
+impl<MessageType: Debug, Service: crate::service::Service> Drop for Sample<MessageType, Service> {
     fn drop(&mut self) {
-        self.subscriber
-            .release_sample(self.channel_id, self.ptr.as_ptr() as usize);
+        match self.publisher_connections.get(self.channel_id) {
+            Some(c) => {
+                let distance =
+                    self.ptr.as_ptr() as usize - c.data_segment.allocator_data_start_address();
+                match c.receiver.release(PointerOffset::new(distance)) {
+                    Ok(()) => (),
+                    Err(ZeroCopyReleaseError::RetrieveBufferFull) => {
+                        fatal_panic!(from self, when c.receiver.release(PointerOffset::new(distance)),
+                                    "This should never happen! The publishers retrieve channel is full and the sample cannot be returned.");
+                    }
+                }
+            }
+            None => {
+                warn!(from self, "Unable to release sample since the connection is broken. The sample will be discarded and has to be reclaimed manually by the publisher.");
+            }
+        }
     }
 }
 
-impl<'subscriber, MessageType: Debug> Payload<MessageType> for Sample<'subscriber, MessageType> {
-    fn payload(&self) -> &MessageType {
+impl<MessageType: Debug, Service: crate::service::Service> Sample<MessageType, Service> {
+    /// Returns a reference to the payload of the sample
+    pub fn payload(&self) -> &MessageType {
         self.ptr.as_data_ref()
     }
 
-    fn header(&self) -> &Header {
+    /// Returns a reference to the header of the sample.
+    pub fn header(&self) -> &Header {
         self.ptr.as_header_ref()
     }
 }
