+++ iceoryx2/src/service/port_factory/subscriber.rs
@@ -33,7 +33,10 @@ use std::fmt::Debug;
 
 use iceoryx2_bb_log::fail;
 
-use crate::{port::subscribe::SubscriberCreateError, port::subscriber::Subscriber, service};
+use crate::{
+    port::subscriber::{Subscriber, SubscriberCreateError},
+    service,
+};
 
 use super::publish_subscribe::PortFactory;
